Method/qwen/requirements.txt
accelerate>=0.30.0
Pillow>=10.0.0
torch>=2.1.0
transformers>=4.57.0
